# GGRC-2323 Improve performance of generating daily digest emails