@plamut 2017-05-31
@Smotko As refactoring would take too much time at this point, and the QA needs this improvement sooner rather than later, I opted to remove the TODO comment, and created a [refactoring ticket](https://gojira.corp.google.com/browse/GGRC-2342) instead.

Please let me know if there is anything to change in its description to make it more clear.