@plamut 2017-06-05
@zidarsk8 Created a [ticket](https://gojira.corp.google.com/browse/GGRC-2368). Its description is very short, but I believe a link to this PR will suffice (since it's highly unlikely the link would become dead).